fix : sign_up 스타일 컴포넌트, 페이지 컴포넌트로 머지
